Move config tests after production items
